refactor(studio): deduplicate bench helper ownership
