+++ Automattic/studio/bench/lib/wordpress-quality.mjs
@@ -1,7 +1,15 @@
-import { mkdir, readFile, readdir, stat, writeFile } from 'node:fs/promises';
-import path from 'node:path';
+import { readFile } from 'node:fs/promises';
 
 import { runCli as defaultRunCli } from './studio-bench.mjs';
+import { collectLatestGeneratedTheme } from './design-gates.mjs';
+
+export {
+  collectGeneratedThemeUxGates,
+  collectThemeBlockDocuments,
+  hiddenEditorContentDiagnostics,
+  reportedFreeformBlockCount,
+  structuralSelectorDriftDiagnostics,
+} from './design-gates.mjs';
 
 export const QUALITY_PROBE = String.raw`
 function bench_count_blocks( $blocks, &$counts ) {
@@ -156,76 +164,6 @@ function parseQualityProbeOutput(stdout) {
   return JSON.parse(stdout.slice(jsonStart));
 }
 
-export async function collectThemeBlockDocuments(sitePath) {
-  const { themeRoot } = await collectLatestGeneratedTheme(sitePath);
-  if (!themeRoot) {
-    return [];
-  }
-
-  const documents = [];
-
-  async function collectDir(relativeDir, extension) {
-    const dir = path.join(themeRoot, relativeDir);
-    let entries = [];
-    try {
-      entries = await readdir(dir, { withFileTypes: true });
-    } catch {
-      return;
-    }
-
-    for (const entry of entries) {
-      if (!entry.isFile() || !entry.name.endsWith(extension)) {
-        continue;
-      }
-      const file = path.join(dir, entry.name);
-      let content = await readFile(file, 'utf8');
-      if (extension === '.php') {
-        content = content.replace(/^\s*<\?php[\s\S]*?\?>\s*/, '');
-      }
-      if (content.includes('<!-- wp:')) {
-        documents.push({
-          source: path.relative(themeRoot, file),
-          content,
-        });
-      }
-    }
-  }
-
-  await collectDir('templates', '.html');
-  await collectDir('parts', '.html');
-  await collectDir('patterns', '.php');
-
-  return documents;
-}
-
-export async function collectLatestGeneratedTheme(sitePath) {
-  const themesRoot = path.join(sitePath, 'wp-content/themes');
-  let themeEntries = [];
-  try {
-    themeEntries = await readdir(themesRoot, { withFileTypes: true });
-  } catch {
-    return { themeRoot: '', themeSlug: '', reportPath: '', error: 'Themes directory not found.' };
-  }
-
-  const generatedThemes = [];
-  for (const entry of themeEntries) {
-    if (!entry.isDirectory()) {
-      continue;
-    }
-    const themeRoot = path.join(themesRoot, entry.name);
-    const reportPath = path.join(themeRoot, 'import-report.json');
-    try {
-      const reportStat = await stat(reportPath);
-      generatedThemes.push({ themeRoot, themeSlug: entry.name, reportPath, mtimeMs: reportStat.mtimeMs });
-    } catch {
-      // Ignore non-importer themes.
-    }
-  }
-
-  generatedThemes.sort((a, b) => b.mtimeMs - a.mtimeMs);
-  return generatedThemes[0] || { themeRoot: '', themeSlug: '', reportPath: '', error: 'No Static Site Importer report found.' };
-}
-
 export async function collectLatestImportReport(sitePath) {
   const { reportPath, error } = await collectLatestGeneratedTheme(sitePath);
   if (!reportPath) {
@@ -286,464 +224,6 @@ export function importerTimingMetrics(importReport) {
   return metrics;
 }
 
-async function readIfExists(file) {
-  try {
-    return await readFile(file, 'utf8');
-  } catch {
-    return '';
-  }
-}
-
-async function collectCssFiles(root, maxFiles = 40) {
-  const files = [];
-
-  async function visit(dir) {
-    if (files.length >= maxFiles) {
-      return;
-    }
-    let entries = [];
-    try {
-      entries = await readdir(dir, { withFileTypes: true });
-    } catch {
-      return;
-    }
-
-    for (const entry of entries) {
-      if (files.length >= maxFiles) {
-        return;
-      }
-      if (entry.name === 'node_modules' || entry.name === '.git') {
-        continue;
-      }
-      const absolute = path.join(dir, entry.name);
-      if (entry.isDirectory()) {
-        await visit(absolute);
-      } else if (entry.isFile() && entry.name.endsWith('.css')) {
-        files.push(absolute);
-      }
-    }
-  }
-
-  await visit(root);
-  return files;
-}
-
-export function reportedFreeformBlockCount(importReport) {
-  const quality = importReport?.report?.quality || {};
-  const generatedTheme = importReport?.report?.generated_theme || {};
-  const candidates = [
-    quality.freeform_block_count,
-    quality.freeform_blocks,
-    generatedTheme.freeform_block_count,
-    generatedTheme.freeform_blocks,
-  ];
-
-  for (const candidate of candidates) {
-    const value = Number(candidate);
-    if (Number.isFinite(value)) {
-      return value;
-    }
-  }
-
-  return 0;
-}
-
-function freeformDiagnostics(documents) {
-  const results = [];
-  let count = 0;
-
-  for (const document of documents) {
-    const documentCount = countRegex(document.content, /<!--\s+wp:freeform\b/gi);
-    if (documentCount === 0) {
-      continue;
-    }
-    count += documentCount;
-    results.push({
-      source: document.source,
-      freeform_block_count: documentCount,
-    });
-  }
-
-  return { count, results };
-}
-
-function splitSelectorList(selector) {
-  const selectors = [];
-  let current = '';
-  let depth = 0;
-  let quote = '';
-
-  for (const char of String(selector || '')) {
-    if (quote) {
-      current += char;
-      if (char === quote) {
-        quote = '';
-      }
-      continue;
-    }
-
-    if (char === '"' || char === "'") {
-      quote = char;
-      current += char;
-      continue;
-    }
-
-    if (char === '(' || char === '[') {
-      depth += 1;
-      current += char;
-      continue;
-    }
-    if (char === ')' || char === ']') {
-      depth = Math.max(0, depth - 1);
-      current += char;
-      continue;
-    }
-
-    if (char === ',' && depth === 0) {
-      const value = current.trim();
-      if (value) {
-        selectors.push(value);
-      }
-      current = '';
-      continue;
-    }
-
-    current += char;
-  }
-
-  const value = current.trim();
-  if (value) {
-    selectors.push(value);
-  }
-
-  return selectors;
-}
-
-function selectorCoverageTokens(selector) {
-  const tokens = new Set();
-  for (const match of String(selector || '').matchAll(/\.(-?[_a-zA-Z]+[_a-zA-Z0-9-]*)/g)) {
-    const name = match[1];
-    if (!/^(?:editor-styles-wrapper|block-editor|wp-admin|is-root-container)$/.test(name)) {
-      tokens.add(`.${name}`);
-    }
-  }
-  for (const match of String(selector || '').matchAll(/\[\s*([^\]\s~|^$*=]+)/g)) {
-    tokens.add(`[${match[1].toLowerCase()}]`);
-  }
-  return tokens;
-}
-
-function editorOverrideCoversHiddenSelector(hiddenSelector, overrideSelector) {
-  const hiddenTokens = selectorCoverageTokens(hiddenSelector);
-  if (hiddenTokens.size === 0) {
-    return false;
-  }
-
-  const overrideTokens = selectorCoverageTokens(overrideSelector);
-  if (overrideTokens.size === 0) {
-    return false;
-  }
-
-  for (const token of overrideTokens) {
-    if (!hiddenTokens.has(token)) {
-      return false;
-    }
-  }
-
-  return true;
-}
-
-export function hiddenEditorContentDiagnostics(cssFiles) {
-  const hiddenRules = [];
-  const editorOverrideRules = [];
-  const hiddenSelectors = [];
-  const editorOverrideSelectors = [];
-  const revealSelectorPattern = /(?:^|[\s.#:[>,])(?:js[-_])?(?:reveal|revealed|aos|fade[-_]in|animate[-_]on[-_]scroll|scroll[-_]reveal)\b|\[data[-_](?:reveal|aos|animate)/i;
-  const hiddenSelectorPattern = /(?:^|[\s.#:[>,])hidden\b/i;
-  const hiddenDeclarationPattern = /(?:opacity\s*:\s*0(?:\.0+)?\b|visibility\s*:\s*hidden\b|display\s*:\s*none\b)/i;
-  const visibleDeclarationPattern = /(?:opacity\s*:\s*1(?:\.0+)?\b|visibility\s*:\s*visible\b|display\s*:\s*(?:block|grid|flex|contents)\b)/i;
-  const editorSelectorPattern = /(?:editor-styles-wrapper|block-editor|wp-admin|is-root-container)/i;
-
-  for (const file of cssFiles) {
-    const content = typeof file.content === 'string' ? file.content : '';
-    for (const match of content.matchAll(/([^{}]+)\{([^{}]*)\}/g)) {
-      const selector = match[1].trim().replace(/\s+/g, ' ');
-      const declarations = match[2].trim().replace(/\s+/g, ' ');
-      const isRevealSelector = revealSelectorPattern.test(selector);
-      const isBroadEditorHiddenOverride =
-        editorSelectorPattern.test(selector) && hiddenSelectorPattern.test(selector) && visibleDeclarationPattern.test(declarations);
-      if (!isRevealSelector && !isBroadEditorHiddenOverride) {
-        continue;
-      }
-
-      const rule = {
-        source: file.source,
-        selector,
-        declarations: declarations.slice(0, 300),
-      };
-
-      const selectorList = splitSelectorList(selector);
-
-      if (isRevealSelector && hiddenDeclarationPattern.test(declarations)) {
-        hiddenRules.push(rule);
-        for (const selectorItem of selectorList) {
-          if (revealSelectorPattern.test(selectorItem)) {
-            hiddenSelectors.push({ ...rule, selector: selectorItem });
-          }
-        }
-      }
-      if (editorSelectorPattern.test(selector) && visibleDeclarationPattern.test(declarations)) {
-        editorOverrideRules.push(rule);
-        for (const selectorItem of selectorList) {
-          if (editorSelectorPattern.test(selectorItem)) {
-            editorOverrideSelectors.push({ ...rule, selector: selectorItem });
-          }
-        }
-      }
-    }
-  }
-
-  const missingEditorOverrideSelectors = hiddenSelectors.filter(
-    (hiddenRule) =>
-      !editorOverrideSelectors.some((editorRule) =>
-        editorOverrideCoversHiddenSelector(hiddenRule.selector, editorRule.selector)
-      )
-  );
-
-  return {
-    hidden_rule_count: hiddenRules.length,
-    editor_override_rule_count: editorOverrideRules.length,
-    missing_editor_override_count: missingEditorOverrideSelectors.length,
-    hidden_rules: hiddenRules.slice(0, 25),
-    editor_override_rules: editorOverrideRules.slice(0, 25),
-    missing_editor_override_rules: missingEditorOverrideSelectors.slice(0, 25),
-  };
-}
-
-function cssRules(cssFiles) {
-  const rules = [];
-  for (const file of cssFiles) {
-    const content = typeof file.content === 'string' ? file.content : '';
-    for (const match of content.matchAll(/([^{}]+)\{([^{}]*)\}/g)) {
-      const selector = match[1].trim().replace(/\s+/g, ' ');
-      if (!selector || selector.startsWith('@')) {
-        continue;
-      }
-      rules.push({
-        source: file.source,
-        selector,
-        declarations: match[2].trim().replace(/\s+/g, ' '),
-      });
-    }
-  }
-  return rules;
-}
-
-function structuralSelectorTarget(selector) {
-  const normalizedSelector = String(selector || '')
-    .replace(/::?[a-z-]+(?:\([^)]*\))?/gi, '')
-    .trim();
-  const match = normalizedSelector.match(
-    /(?:^|[\s>+~])(?<tag>header|main|nav|footer|section|article|aside)(?<compound>(?:[#.][-_a-zA-Z0-9]+)*)/i
-  );
-  if (!match?.groups) {
-    return null;
-  }
-
-  const compound = match.groups.compound || '';
-  const id = compound.match(/#([-_a-zA-Z0-9]+)/)?.[1] || '';
-  const classes = [...compound.matchAll(/\.([-_a-zA-Z0-9]+)/g)].map((classMatch) => classMatch[1]);
-  const text = `${match.groups.tag} ${id} ${classes.join(' ')} ${selector}`;
-  const heroLike = /\bhero\b/i.test(text);
-  if (!id && classes.length === 0) {
-    return null;
-  }
-
-  return {
-    tag: match.groups.tag.toLowerCase(),
-    id,
-    classes,
-    hero_like: heroLike,
-  };
-}
-
-function hasMaterialStructuralLayout(declarations, target) {
-  const layoutPattern = /(?:display\s*:\s*(?:grid|flex)|grid-template|flex(?:-direction|-wrap)?\s*:|align-items\s*:|justify-content\s*:|min-height\s*:|height\s*:\s*(?:\d+(?:vh|dvh|svh)|100%)|padding(?:-[a-z]+)?\s*:|background(?:-|\s*:)|position\s*:\s*(?:relative|absolute|fixed|sticky))/i;
-  return target?.hero_like === true && layoutPattern.test(String(declarations || ''));
-}
-
-function htmlAttributeTokens(attributes, name) {
-  const values = [];
-  const pattern = new RegExp(`\\b${name}\\s*=\\s*(["'])(.*?)\\1`, 'gi');
-  for (const match of String(attributes || '').matchAll(pattern)) {
-    values.push(...String(match[2] || '').split(/\s+/).filter(Boolean));
-  }
-  return values;
-}
-
-function markupHasId(markup, id) {
-  if (!id) {
-    return true;
-  }
-  const escaped = id.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-  return (
-    new RegExp(`\\bid\\s*=\\s*(["'])${escaped}\\1`, 'i').test(markup) ||
-    new RegExp(`"anchor"\\s*:\\s*"${escaped}"`, 'i').test(markup)
-  );
-}
-
-function markupHasClasses(markup, classes) {
-  const classValues = [];
-  const pattern = /\bclass(?:Name)?\s*=\s*(["'])(.*?)\1/gi;
-  for (const match of String(markup || '').matchAll(pattern)) {
-    classValues.push(...String(match[2] || '').split(/\s+/).filter(Boolean));
-  }
-  return classes.every((className) => classValues.includes(className));
-}
-
-function markupHasStructuralTarget(markup, target) {
-  const tagPattern = new RegExp(`<${target.tag}\\b([^>]*)>`, 'gi');
-  for (const match of String(markup || '').matchAll(tagPattern)) {
-    const attributes = match[1] || '';
-    const ids = htmlAttributeTokens(attributes, 'id');
-    const classes = htmlAttributeTokens(attributes, 'class');
-    const idMatches = !target.id || ids.includes(target.id);
-    const classesMatch = target.classes.every((className) => classes.includes(className));
-    if (idMatches && classesMatch) {
-      return true;
-    }
-  }
-  return false;
-}
-
-function structuralSelectorDriftReason(markup, target) {
-  if (target.id && !markupHasId(markup, target.id)) {
-    return 'missing_generated_dom_id';
-  }
-  if (target.classes.length > 0 && !markupHasClasses(markup, target.classes)) {
-    return 'missing_generated_dom_class';
-  }
-  return 'generated_dom_tag_drift';
-}
-
-export function structuralSelectorDriftDiagnostics(cssFiles, generatedDocuments) {
-  const generatedMarkup = generatedDocuments.map((document) => document.content || '').join('\n');
-  const materialSelectors = [];
-  const missingSelectors = [];
-
-  for (const rule of cssRules(cssFiles)) {
-    for (const selector of splitSelectorList(rule.selector)) {
-      const target = structuralSelectorTarget(selector);
-      if (!target || !hasMaterialStructuralLayout(rule.declarations, target)) {
-        continue;
-      }
-
-      const diagnostic = {
-        source: rule.source,
-        selector,
-        expected_tag: target.tag,
-        expected_id: target.id,
-        expected_classes: target.classes,
-        declarations: rule.declarations.slice(0, 300),
-      };
-      materialSelectors.push(diagnostic);
-      if (!markupHasStructuralTarget(generatedMarkup, target)) {
-        missingSelectors.push({
-          ...diagnostic,
-          reason: structuralSelectorDriftReason(generatedMarkup, target),
-        });
-      }
-    }
-  }
-
-  return {
-    source_structural_selector_count: materialSelectors.length,
-    missing_structural_selector_count: missingSelectors.length,
-    source_structural_selectors: materialSelectors.slice(0, 25),
-    missing_structural_selectors: missingSelectors.slice(0, 25),
-  };
-}
-
-export async function collectGeneratedThemeUxGates(sitePath, importReport, artifactDir) {
-  const { themeRoot, themeSlug, error } = await collectLatestGeneratedTheme(sitePath);
-  const artifactPath = path.join(artifactDir, 'generated-theme-ux-gates.json');
-  const reasons = [];
-
-  if (!themeRoot) {
-    const skipped = {
-      skipped: true,
-      error,
-      theme_slug: themeSlug,
-      artifact: artifactPath,
-      generated_theme_ux_quality_pass: false,
-      generated_theme_ux_quality_failure_count: 1,
-      generated_theme_ux_quality_failure_reasons: ['missing_generated_theme'],
-    };
-    await writeFile(artifactPath, JSON.stringify(skipped, null, 2));
-    return skipped;
-  }
-
-  const documents = await collectThemeBlockDocuments(sitePath);
-  const freeform = freeformDiagnostics(documents);
-  const importerFreeformCount = reportedFreeformBlockCount(importReport);
-  const cssFiles = await collectCssFiles(themeRoot);
-  const cssFileContents = [];
-  for (const file of cssFiles) {
-    cssFileContents.push({ source: path.relative(themeRoot, file), content: await readIfExists(file) });
-  }
-  const hiddenEditorContent = hiddenEditorContentDiagnostics(cssFileContents);
-  const sourceRoot = path.join(sitePath, 'tmp/static-site');
-  const sourceCssFiles = await collectCssFiles(sourceRoot);
-  const sourceCssFileContents = [];
-  for (const file of sourceCssFiles) {
-    sourceCssFileContents.push({ source: path.relative(sourceRoot, file), content: await readIfExists(file) });
-  }
-  const structuralSelectorDrift = structuralSelectorDriftDiagnostics(sourceCssFileContents, documents);
-
-  if (freeform.count !== importerFreeformCount) {
-    reasons.push('freeform_report_count_mismatch');
-  }
-  if (hiddenEditorContent.missing_editor_override_count > 0) {
-    reasons.push('css_hidden_editor_content_without_override');
-  }
-  if (structuralSelectorDrift.missing_structural_selector_count > 0) {
-    reasons.push('source_css_structural_selector_drift');
-  }
-
-  const diagnostics = {
-    skipped: false,
-    theme_slug: themeSlug,
-    theme_root: themeRoot,
-    document_count: documents.length,
-    css_file_count: cssFiles.length,
-    actual_freeform_block_count: freeform.count,
-    importer_freeform_block_count: importerFreeformCount,
-    freeform_report_mismatch_count: freeform.count === importerFreeformCount ? 0 : 1,
-    freeform_documents: freeform.results,
-    css_hidden_editor_content_count: hiddenEditorContent.hidden_rule_count,
-    css_editor_reveal_override_count: hiddenEditorContent.editor_override_rule_count,
-    css_hidden_editor_content_without_override_count: hiddenEditorContent.missing_editor_override_count,
-    css_hidden_editor_content: hiddenEditorContent.hidden_rules,
-    css_editor_reveal_overrides: hiddenEditorContent.editor_override_rules,
-    source_css_file_count: sourceCssFiles.length,
-    source_css_structural_selector_count: structuralSelectorDrift.source_structural_selector_count,
-    source_css_missing_structural_selector_count: structuralSelectorDrift.missing_structural_selector_count,
-    source_css_missing_structural_selectors: structuralSelectorDrift.missing_structural_selectors,
-    generated_theme_ux_quality_pass: reasons.length === 0,
-    generated_theme_ux_quality_failure_count: reasons.length,
-    generated_theme_ux_quality_failure_reasons: reasons,
-    remaining_manual_gates: [
-      'site_editor_canvas_above_fold_visible_text',
-      'footer_utility_links_not_responsive_nav_overlay',
-      'admin_bar_not_obscured_by_fixed_or_sticky_chrome',
-    ],
-    artifact: artifactPath,
-  };
-
-  await writeFile(artifactPath, JSON.stringify(diagnostics, null, 2));
-  return diagnostics;
-}
-
 export function agentAuthoredBlockMetrics(result) {
   const toolCalls = Array.isArray(result?.toolCalls) ? result.toolCalls : [];
   const writeCalls = toolCalls.filter((item) => item && item.name === 'Write');
@@ -783,7 +263,6 @@ export function nativeBlockQualityMetrics(quality, authoredBlocks, editorValidat
   if (targetPagesSeen === 0 || targetPostsWithBlocks === 0) {
     reasons.push('missing_target_block_page');
   }
-
   if (agentAuthoredWpHtmlOpeners > 0) {
     reasons.push('agent_authored_wp_html');
   }
