+++ Automattic/studio/bench/studio-agent-site-build.bench.mjs
@@ -37,10 +37,29 @@ export { compareVisualFidelity } from './lib/visual-fidelity.mjs';
 
 import {
   compareSemanticFidelity as compareSemanticFidelityImpl,
-  semanticMismatchFailureDetails,
   semanticTargetMetric,
 } from './lib/semantic-fidelity.mjs';
 export { semanticMismatchFailureDetails, semanticTargetMetric } from './lib/semantic-fidelity.mjs';
+
+import {
+  agentSuccessGate,
+  importerBlockQualityMetrics,
+  importerBlockQualityFailureDetails,
+  VISUAL_PIXEL_DIFF_THRESHOLD,
+  visualEditorParityMetrics,
+  visualEditorParityFailureDetails,
+  visualPixelDiffFailureDetails,
+} from './lib/site-build-gates.mjs';
+
+export {
+  agentSuccessGate,
+  importerBlockQualityMetrics,
+  importerBlockQualityFailureDetails,
+  VISUAL_PIXEL_DIFF_THRESHOLD,
+  visualEditorParityMetrics,
+  visualEditorParityFailureDetails,
+  visualPixelDiffFailureDetails,
+} from './lib/site-build-gates.mjs';
 
 const STUDIO_PATH = process.env.HOMEBOY_COMPONENT_PATH;
 const SHARED_STATE = process.env.HOMEBOY_BENCH_SHARED_STATE || os.tmpdir();
@@ -57,11 +76,6 @@ const DEFAULT_PROMPT_VARIANT = 'studio-code';
 const PROMPT_CATEGORY = 'site-build';
 const SYSTEM_PROMPT_FILES = ['apps/cli/ai/system-prompt.ts'];
 const requireFromBench = createRequire(import.meta.url);
-const VISUAL_EDITOR_PIXEL_DIFF_THRESHOLD = 0.05;
-// Whole-page screenshots have minor browser/font antialiasing noise; 5% keeps
-// that noise green while failing visible regressions such as missing icons.
-export const VISUAL_PIXEL_DIFF_THRESHOLD = 0.05;
-
 if (!STUDIO_PATH) {
   throw new Error('HOMEBOY_COMPONENT_PATH is required');
 }
@@ -841,142 +855,6 @@ function optionalArtifactPath(name, value) {
   return typeof value === 'string' && value.length > 0 ? { [name]: value } : {};
 }
 
-
-export function importerBlockQualityMetrics(importReport) {
-  const quality = importReport?.report?.quality || {};
-
-  return {
-    importerCoreHtmlBlockCount: metric(quality.core_html_block_count),
-    importerFreeformBlockCount: metric(quality.freeform_block_count),
-    importerFallbackCount: metric(quality.fallback_count),
-  };
-}
-
-export function importerBlockQualityFailureDetails(importerBlockQuality) {
-  const { importerCoreHtmlBlockCount, importerFreeformBlockCount, importerFallbackCount } = importerBlockQuality;
-
-  if (importerCoreHtmlBlockCount === 0 && importerFreeformBlockCount === 0 && importerFallbackCount === 0) {
-    return [];
-  }
-
-  return [
-    `importer block quality: core/html=${importerCoreHtmlBlockCount}, freeform=${importerFreeformBlockCount}, fallback=${importerFallbackCount}`,
-  ];
-}
-
-function visualRatio(value) {
-  const ratio = metric(value);
-  return Number.isFinite(ratio) ? ratio : 1;
-}
-
-function formatVisualRatio(value) {
-  return visualRatio(value).toFixed(2);
-}
-
-export function visualEditorParityMetrics(visualComparison) {
-  return {
-    visualEditorVsSourcePixelDiffRatio: visualRatio(visualComparison?.visual_editor_vs_source_pixel_diff_ratio),
-    visualEditorVsFrontendPixelDiffRatio: visualRatio(visualComparison?.visual_editor_vs_frontend_pixel_diff_ratio),
-    visualSourceVsFrontendPixelDiffRatio: visualRatio(
-      visualComparison?.visual_pixel_diff_ratio ??
-        visualComparison?.pixel_diff_ratio ??
-        visualComparison?.visual_source_vs_frontend_pixel_diff_ratio_diagnostic
-    ),
-    visualEditorParityErrorCount: metric(visualComparison?.visual_editor_parity_error_count),
-  };
-}
-
-export function visualEditorParityFailureDetails(visualEditorParity) {
-  const {
-    visualEditorVsSourcePixelDiffRatio,
-    visualEditorVsFrontendPixelDiffRatio,
-    visualSourceVsFrontendPixelDiffRatio,
-    visualEditorParityErrorCount,
-  } = visualEditorParity;
-  const editorFailedSource = visualEditorVsSourcePixelDiffRatio > VISUAL_EDITOR_PIXEL_DIFF_THRESHOLD;
-  const editorFailedFrontend = visualEditorVsFrontendPixelDiffRatio > VISUAL_EDITOR_PIXEL_DIFF_THRESHOLD;
-
-  if (visualEditorParityErrorCount > 0) {
-    return [`editor visual parity could not be measured (${visualEditorParityErrorCount} capture/diff errors)`];
-  }
-
-  if (!editorFailedSource && !editorFailedFrontend) {
-    return [];
-  }
-
-  if (editorFailedFrontend && visualSourceVsFrontendPixelDiffRatio <= VISUAL_EDITOR_PIXEL_DIFF_THRESHOLD) {
-    return [
-      `editor render diverges from frontend (editor diff: ${formatVisualRatio(
-        visualEditorVsSourcePixelDiffRatio
-      )}, frontend diff: ${formatVisualRatio(
-        visualSourceVsFrontendPixelDiffRatio
-      )}) - likely block-validation or unscoped CSS`,
-    ];
-  }
-
-  if (editorFailedSource && visualSourceVsFrontendPixelDiffRatio > VISUAL_EDITOR_PIXEL_DIFF_THRESHOLD) {
-    return [
-      `editor and frontend both diverge from source (editor: ${formatVisualRatio(
-        visualEditorVsSourcePixelDiffRatio
-      )}, frontend: ${formatVisualRatio(visualSourceVsFrontendPixelDiffRatio)}) - conversion failed before editor concern`,
-    ];
-  }
-
-  return [
-    `editor visual parity failed (editor vs source: ${formatVisualRatio(
-      visualEditorVsSourcePixelDiffRatio
-    )}, editor vs frontend: ${formatVisualRatio(visualEditorVsFrontendPixelDiffRatio)})`,
-  ];
-}
-
-export function visualPixelDiffFailureDetails(visualComparison) {
-  const visualPixelDiffRatio = metric(visualComparison?.pixel_diff_ratio);
-  if (visualPixelDiffRatio <= VISUAL_PIXEL_DIFF_THRESHOLD) {
-    return [];
-  }
-
-  return [
-    `visual pixel diff: ${visualPixelDiffRatio.toFixed(3)} (threshold: ${VISUAL_PIXEL_DIFF_THRESHOLD.toFixed(3)})`,
-  ];
-}
-
-export function agentSuccessGate(result, semanticComparison, importReport, visualComparison) {
-  const semanticMismatchCount = metric(semanticComparison?.mismatch_count);
-  const importerBlockQuality = importerBlockQualityMetrics(importReport);
-  const visualEditorParity = visualEditorParityMetrics(visualComparison);
-  const visualPixelDiffRatio = metric(visualComparison?.pixel_diff_ratio);
-  const agentTimedOut = result?.timedOut === true;
-  const agentSucceeded =
-    result?.success === true &&
-    !result?.error &&
-    !agentTimedOut &&
-    semanticMismatchCount === 0 &&
-    importerBlockQuality.importerCoreHtmlBlockCount === 0 &&
-    importerBlockQuality.importerFreeformBlockCount === 0 &&
-    importerBlockQuality.importerFallbackCount === 0 &&
-    visualEditorParity.visualEditorParityErrorCount === 0 &&
-    visualEditorParity.visualEditorVsSourcePixelDiffRatio <= VISUAL_EDITOR_PIXEL_DIFF_THRESHOLD &&
-    visualEditorParity.visualEditorVsFrontendPixelDiffRatio <= VISUAL_EDITOR_PIXEL_DIFF_THRESHOLD &&
-    visualPixelDiffRatio <= VISUAL_PIXEL_DIFF_THRESHOLD;
-
-  return {
-    agentSucceeded,
-    semanticMismatchCount,
-    semanticFailureDetails: semanticMismatchCount > 0 ? semanticMismatchFailureDetails(semanticComparison) : [],
-    importerBlockQuality,
-    importerBlockQualityFailureDetails: importerBlockQualityFailureDetails(importerBlockQuality),
-    visualEditorParity,
-    visualEditorFailureDetails: visualEditorParityFailureDetails(visualEditorParity),
-    visualPixelDiffRatio,
-    visualPixelDiffFailureDetails: visualPixelDiffFailureDetails(visualComparison),
-    metrics: {
-      success_rate: agentSucceeded ? 1 : 0,
-      agent_error_rate: agentSucceeded ? 0 : 1,
-      timed_out: agentTimedOut ? 1 : 0,
-      agent_runner_error: typeof result?.error === 'string' ? 1 : 0,
-    },
-  };
-}
 
 export default async function studioAgentSiteBuildBench() {
   const runtime = resolveBenchRuntime();
